% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_example.R
\name{bar_example}
\alias{bar_example}
\title{A Static Bar Chart example}
\usage{
bar_example()
}
\description{
This is just a simple example about how to have a function that starts to create a bar chart and label it.
This chart is inspired in the tutorial available here - http://tryr.codeschool.com/levels/2/challenges/1
}
\examples{
bar_example()
}
\keyword{barchart,}
\keyword{legend,}
\keyword{pirates}
\keyword{vector,}

