% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{decision_tree}
\alias{decision_tree}
\title{Dynamic function to return a decision tree}
\usage{
decision_tree(formula, data_table)
}
\description{
Based on data sent from Qlik Sense this function will call rpart and plot it's result.
}
\examples{
decision_tree(formula, data_table)
}
\keyword{decision,}
\keyword{plot,}
\keyword{qlik,}
\keyword{rpart,}
\keyword{sense}
\keyword{tree,}

