% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ts.R
\name{forecast_ts}
\alias{forecast_ts}
\title{Dynamic function to return time series forecasting plot}
\usage{
forecast_ts(datos)
}
\description{
Based on data sent from Qlik Sense this function will call ts() from R's forecast library and plot it based on ets().
}
\examples{
forecast_ts(data_table)
}
\keyword{ets,}
\keyword{forecast,}
\keyword{plot,}
\keyword{qlik,}
\keyword{sense}
\keyword{series,}
\keyword{time,}
\keyword{ts,}

