% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_function.R
\name{hello_function}
\alias{hello_function}
\title{An Hello Wolrd function}
\usage{
hello_function()
}
\description{
This is just a sample of an hello wolrd function to say hi!
}
\examples{
hello_function()
}
\keyword{hello}

