% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_example.R
\name{volcano_example}
\alias{volcano_example}
\title{A static perspective example for a 3D Volcano}
\usage{
volcano_example()
}
\description{
This is an example to show 3 Dimensional plot. The data is a sample R data of a dormant New Zealand Volcano.
This is inspired in the tutorial available here - http://tryr.codeschool.com/levels/3/challenges/1
}
\examples{
volcano_example()
}
\keyword{3d,}
\keyword{advanced}
\keyword{matrix,}
\keyword{plot,}

